// src/app/shared/models/api-response/receipt-shipment-response.ts
export class ReceiptShipmentResponse {
    departmentName: string;
    productName: string;
    stockCardShipmentItemProductQty: number;
    productGroupName: string;
    stockCardOpeningQty: number;
    measureUnitName: string;
    stockCardReceiptItemProductQty: number;
    stockCardEndQty: number;
}