// src/app/shared/models/phuquoc/receipt-shipment.ts
export class ReceiptShipment {
    description: string;
    measureUnitName: string;
    stockCardShipmentItemProductQty: number;
    stockCardOpeningBalanceAmount: number;
    stockCardReceiptItemProductQty: number;
    stockCardEndBalanceAmount: number;
    type: string;
    show: boolean;
    collapse: boolean;
    collapseGroup: boolean;
    department: string;
    groupName: string;
}